part3_a_3.5 in progress
I have started implementing a post method, however I want to have my ids properly done rather then assigning a random ones an hope they do not match. It makes things a bit more complicated. I sort of got a pseudocode for what I want to do, which is make a set of ids taken and subtract it from the indices  of persons list, and if there is a difference grater than an empty set, then I'm going to use it for ids. The code needs to be fixed thought, because it is sorta pseudo code as I've mentioned before.